requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile pyproject.toml
#
annotated-types==0.6.0
    # via pydantic
attrs==23.2.0
    # via maus
click==8.1.7
    # via kohlrahbi (pyproject.toml)
colorlog==6.8.2
    # via kohlrahbi (pyproject.toml)
et-xmlfile==1.1.0
    # via openpyxl
lxml==5.1.1
    # via python-docx
marshmallow==3.21.1
    # via maus
maus==0.4.2
    # via kohlrahbi (pyproject.toml)
more-itertools==10.2.0
    # via maus
numpy==1.26.4
    # via pandas
openpyxl==3.1.2
    # via kohlrahbi (pyproject.toml)
packaging==24.0
    # via marshmallow
pandas==2.2.1
    # via kohlrahbi (pyproject.toml)
pydantic==2.6.4
    # via kohlrahbi (pyproject.toml)
pydantic-core==2.16.3
    # via pydantic
python-dateutil==2.9.0.post0
    # via pandas
python-docx==1.1.0
    # via kohlrahbi (pyproject.toml)
pytz==2024.1
    # via
    #   kohlrahbi (pyproject.toml)
    #   pandas
six==1.16.0
    # via python-dateutil
tomlkit==0.12.4
    # via kohlrahbi (pyproject.toml)
typing-extensions==4.10.0
    # via
    #   pydantic
    #   pydantic-core
    #   python-docx
tzdata==2024.1
    # via pandas
xlsxwriter==3.2.0
    # via kohlrahbi (pyproject.toml)
